backend/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
supabase
pydantic>=2.0
pydantic-settings
python-dotenv
python-multipart
reportlab
httpx
pypdf
tiktoken
openai
pandas
openpyxl
python-docx
tabulate